requirements.txt
aiosmtplib==2.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
Authlib==0.15.5
authx==0.9.1
bcrypt==4.0.1
bidict==0.22.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.0
databases==0.7.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.95.2
fastapi-sqla==2.9.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
motor==3.1.2
mysql==0.0.3
mysqlclient==2.2.0
passlib==1.7.4
prometheus-client==0.17.1
pycparser==2.21
pydantic==1.10.7
pydantic_core==2.6.1
pyinstrument==4.4.0
PyJWT==2.7.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.6.0
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
redis==4.5.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
structlog==23.1.0
typing_extensions==4.7.1
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
Pillow==10.0.1
python-multipart==0.0.6
pymysql==1.1.0
aioshutil==1.3
fleep==1.0.1